Add files via upload
-Briefly summarize the DriverPass project. Who was the client? What type of system did they want you to design?
The client (DriverPass) saw an opportunity for training at their local DMV due to the number of people who fail their driving test. To address that opportunity the client wanted a system created to offer better driver training to anyone preparing for an upcoming driving test. The client wanted to do this by the system offering on-the-road training, learning resources, and practice tests.

-What did you do particularly well?
I believe I did well determining the requirements of the system from the interview transcript and developing those requirements into objectives and goals.

-If you could choose one part of your work on these documents to revise, what would you pick? How would you improve it?
I believe that I would revise the platform constraints and limitations to incorporate more elements. I feel the current version does not offer enough detail in these aspects and could use more research to determine further constraints and limitations.

-How did you interpret the user’s needs and implement them into your system design? Why is it so important to consider the user’s needs when designing?
I interpreted the users' needs presented in the interview transcript into requirements. These needs included both functional and non-functional requirements that defined what the DriverPass team envisioned for their system. Overall these requirements were translated into the design of the system which turned into the system that the clients had envisioned and fulfilled the opportunity presented for the end-users.

-How do you approach designing software? What techniques or strategies would you use in the future to analyze and design a system?
To design software, an analyst must meet with key stakeholders such as the client and possible end-users to develop requirements for the system. Based on those requirements the team would create UML diagrams to model and design the systems structure and data flows.